ci: :bug: Report, Store, Order Sheet bug fixes
